inital refactor of schema

+++ app/src/main/java/com/example/myapplication/ReviewModel.java
@@ -6,13 +6,13 @@ import java.util.ArrayList;
 
 public class ReviewModel {
     String reviewTitle;
-    int reviewRating;
+    float reviewRating;
     ArrayList<Bitmap> reviewImages;
     String reviewText;
     int reviewerId;
     int businessId;
 
-    public ReviewModel(int reviewRating, String reviewTitle, ArrayList<Bitmap> reviewImages, String reviewText, int reviewerId, int businessId) {
+    public ReviewModel(float reviewRating, String reviewTitle, ArrayList<Bitmap> reviewImages, String reviewText, int reviewerId, int businessId) {
         this.reviewRating = reviewRating;
         this.reviewTitle = reviewTitle;
         this.reviewImages = reviewImages;
@@ -25,7 +25,7 @@ public class ReviewModel {
         return reviewTitle;
     }
 
-    public int getReviewRating() {
+    public float getReviewRating() {
         return reviewRating;
     }
 
@@ -44,4 +44,5 @@ public class ReviewModel {
     public int getBusinessId() {
         return businessId;
     }
+    
 }
